refactor(course): group pages into subfolders (main/edit/management/settings/import/widgets)
- remove part/part of, make CoursePageTopBar/SwipePageView/VacationView/NoScheduleView/preview/actions standalone with explicit Duration/controller injection
- unify vacation logic in CoursePageController calendarNextSemester with _disposed guard
- move widgets/course/* to pages/course/widgets/* with course_widgets barrel and shim
- extract jwxt_parser and prompt_new_schedule, move import page to import/
- organize by page: main(8)/edit(1)/management(2)/settings(2)/import(2)/widgets(9), drop top-level shims, update external imports
