+++ lib/pages/course/main/course_page_actions.dart
@@ -1,9 +1,27 @@
-part of 'course_page.dart';
+import 'package:flutter/material.dart';
+import 'package:bugaoshan/l10n/app_localizations.dart';
+import 'package:bugaoshan/models/course.dart';
+import '../edit/course_edit_page.dart';
+import '../import/import_schedule_page.dart';
+import 'package:bugaoshan/providers/course_provider.dart';
+import '../widgets/course_detail_sheet.dart';
+import '../widgets/special_day_sheet.dart';
+import 'package:bugaoshan/utils/export_schedule_utils.dart';
+import 'package:bugaoshan/utils/holiday_utils.dart';
+import 'package:bugaoshan/widgets/dialog/dialog.dart';
+import 'package:bugaoshan/widgets/route/router_utils.dart';
+import 'package:bugaoshan/theme_shape.dart';
 
-extension _CoursePageActions on _CoursePageState {
-  void _onImport() {
+/// CoursePage 相关的用户操作集合。
+/// 从原先 `extension _CoursePageActions on _CoursePageState` 抽离为独立 helper，
+/// 通过显式参数注入 [BuildContext] 与 [CourseProvider]，避免持有 State 私有成员。
+class CoursePageActions {
+  static void showImportSheet(
+    BuildContext context,
+    CourseProvider courseProvider,
+  ) {
     final l10n = AppLocalizations.of(context)!;
-    final outerContext = context; // Capture the stable context
+    final outerContext = context;
     showModalBottomSheet(
       context: context,
       shape: const RoundedRectangleBorder(
@@ -84,15 +102,19 @@ extension _CoursePageActions on _CoursePageState {
     );
   }
 
-  void _onExport() {
+  static void showExportSheet(BuildContext context) {
     showExportScheduleSheet(context);
   }
 
-  void _onAddCourse() {
+  static void navigateToAddCourse(BuildContext context) {
     popupOrNavigate(context, const CourseEditPage());
   }
 
-  void _onCourseTap(Course course) {
+  static void showCourseDetailSheet(
+    BuildContext context,
+    Course course,
+    CourseProvider courseProvider,
+  ) {
     showModalBottomSheet(
       context: context,
       isScrollControlled: true,
@@ -106,26 +128,33 @@ extension _CoursePageActions on _CoursePageState {
     );
   }
 
-  void _onCourseLongPress(Course course) {
+  static Future<void> handleCourseLongPress(
+    BuildContext context,
+    Course course,
+    CourseProvider courseProvider,
+  ) async {
     final l10n = AppLocalizations.of(context)!;
-    showYesNoDialog(
+    final confirm = await showYesNoDialog(
       title: l10n.deleteCourse,
       content: l10n.deleteCourseConfirm,
-    ).then((confirm) async {
-      if (confirm == true) {
-        await courseProvider.deleteCourse(course.id);
-      }
-    });
+    );
+    if (confirm == true) {
+      await courseProvider.deleteCourse(course.id);
+    }
   }
 
-  void _onEmptyTap(int dayOfWeek, int section) {
+  static void handleEmptyTap(BuildContext context, int dayOfWeek, int section) {
     popupOrNavigate(
       context,
       CourseEditPage(prefillDayOfWeek: dayOfWeek, prefillSection: section),
     );
   }
 
-  void _onSpecialDayTap(DateTime date, SpecialDayInfo info) {
+  static void handleSpecialDayTap(
+    BuildContext context,
+    DateTime date,
+    SpecialDayInfo info,
+  ) {
     showSpecialDaySheet(context, date, info);
   }
 }
