+++ lib/pages/course/main/course_page_vacation_view.dart
@@ -1,50 +1,30 @@
-part of 'course_page.dart';
+import 'package:bugaoshan/widgets/common/third_center.dart';
+import 'package:flutter/material.dart';
+import 'package:bugaoshan/l10n/app_localizations.dart';
+import 'package:bugaoshan/models/academic_calendar.dart';
+import 'course_page_controller.dart';
+import 'package:bugaoshan/theme_shape.dart';
 
-class _VacationView extends StatefulWidget {
-  final ScheduleConfig scheduleConfig;
-  final List<ScheduleConfig> allSchedules;
+class VacationView extends StatefulWidget {
+  final CoursePageController controller;
   final void Function(AcademicCalendarSemester semester)? onViewNextSemester;
 
-  const _VacationView({
-    required this.scheduleConfig,
-    required this.allSchedules,
+  const VacationView({
+    super.key,
+    required this.controller,
     this.onViewNextSemester,
   });
 
   @override
-  State<_VacationView> createState() => _VacationViewState();
+  State<VacationView> createState() => _VacationViewState();
 }
 
-class _VacationViewState extends State<_VacationView> {
-  AcademicCalendarSemester? _nextSemester;
-  bool _loading = true;
-  bool _hasNextSemesterSchedule = false;
-
+class _VacationViewState extends State<VacationView> {
   @override
   void initState() {
     super.initState();
-    _loadNextSemester();
-  }
-
-  Future<void> _loadNextSemester() async {
-    try {
-      final data = await AcademicCalendarService.loadBundledCalendar();
-      if (mounted) {
-        final next = data.findNextSemester(
-          widget.scheduleConfig.semesterEndDate,
-        );
-        setState(() {
-          _nextSemester = next;
-          _hasNextSemesterSchedule =
-              next != null &&
-              next.findMatchingScheduleId(widget.allSchedules) != null;
-          _loading = false;
-        });
-      }
-    } catch (e) {
-      debugPrint('VacationView: failed to load next semester data: $e');
-      if (mounted) setState(() => _loading = false);
-    }
+    // 触发 controller 懒加载校历，下学期数据由 controller 统一缓存
+    widget.controller.ensureCalendarNextSemester();
   }
 
   @override
@@ -53,11 +33,10 @@ class _VacationViewState extends State<_VacationView> {
     final colorScheme = Theme.of(context).colorScheme;
     final textTheme = Theme.of(context).textTheme;
 
+    final config = widget.controller.config;
     final now = DateTime.now();
     final today = DateTime(now.year, now.month, now.day);
-    final vacationStart = widget.scheduleConfig.semesterEndDate.add(
-      const Duration(days: 1),
-    );
+    final vacationStart = config.semesterEndDate.add(const Duration(days: 1));
     final isOnVacation = !today.isBefore(vacationStart);
 
     return ThirdCenter(
@@ -76,21 +55,34 @@ class _VacationViewState extends State<_VacationView> {
                 textAlign: TextAlign.center,
               ),
               const SizedBox(height: 12),
-              if (_loading)
-                const SizedBox(
-                  width: 20,
-                  height: 20,
-                  child: CircularProgressIndicator(strokeWidth: 2),
-                )
-              else
-                _buildVacationContent(
-                  l10n,
-                  textTheme,
-                  colorScheme,
-                  today,
-                  isOnVacation,
-                  vacationStart,
-                ),
+              ListenableBuilder(
+                listenable: Listenable.merge([
+                  widget.controller.calendarNextSemester,
+                  widget.controller.calendarNextSemesterLoading,
+                ]),
+                builder: (context, _) {
+                  final loading =
+                      widget.controller.calendarNextSemesterLoading.value;
+                  final nextSemester =
+                      widget.controller.calendarNextSemester.value;
+                  if (loading) {
+                    return const SizedBox(
+                      width: 20,
+                      height: 20,
+                      child: CircularProgressIndicator(strokeWidth: 2),
+                    );
+                  }
+                  return _buildVacationContent(
+                    l10n,
+                    textTheme,
+                    colorScheme,
+                    today,
+                    isOnVacation,
+                    vacationStart,
+                    nextSemester,
+                  );
+                },
+              ),
             ],
           ),
         ),
@@ -105,8 +97,9 @@ class _VacationViewState extends State<_VacationView> {
     DateTime today,
     bool isOnVacation,
     DateTime vacationStart,
+    AcademicCalendarSemester? nextSemester,
   ) {
-    if (isOnVacation && _nextSemester == null) {
+    if (isOnVacation && nextSemester == null) {
       return Text(
         l10n.enjoyVacation,
         style: textTheme.bodyLarge?.copyWith(
@@ -117,7 +110,7 @@ class _VacationViewState extends State<_VacationView> {
     }
 
     final daysUntil = isOnVacation
-        ? _nextSemester!.startDate.difference(today).inDays
+        ? nextSemester!.startDate.difference(today).inDays
         : vacationStart.difference(today).inDays;
 
     return Column(
@@ -132,9 +125,9 @@ class _VacationViewState extends State<_VacationView> {
           ),
           textAlign: TextAlign.center,
         ),
-        if (_nextSemester != null) ...[
+        if (nextSemester != null) ...[
           const SizedBox(height: 20),
-          _buildNextSemesterInfo(l10n, textTheme, colorScheme),
+          _buildNextSemesterInfo(l10n, textTheme, colorScheme, nextSemester),
         ],
       ],
     );
@@ -144,9 +137,10 @@ class _VacationViewState extends State<_VacationView> {
     AppLocalizations l10n,
     TextTheme textTheme,
     ColorScheme colorScheme,
+    AcademicCalendarSemester semester,
   ) {
-    final semester = _nextSemester!;
     final regEvent = semester.registrationEvent;
+    final hasSchedule = widget.controller.hasCalendarNextSemesterSchedule;
 
     return Container(
       padding: const EdgeInsets.all(16),
@@ -186,7 +180,7 @@ class _VacationViewState extends State<_VacationView> {
           if (widget.onViewNextSemester != null) ...[
             const SizedBox(height: 12),
             FilledButton.tonal(
-              onPressed: _hasNextSemesterSchedule
+              onPressed: hasSchedule
                   ? () => widget.onViewNextSemester!(semester)
                   : null,
               child: Text(l10n.viewNextSemesterSchedule),
