+++ lib/pages/course/widgets/course_widgets.dart
@@ -1,0 +1,7 @@
+// Barrel for course grid widgets that are shared outside pages/course.
+// External consumers (e.g. class_schedule_inquiry) should import this instead
+// of the internal widgets/course path.
+export 'course_grid.dart';
+export 'course_detail_sheet.dart';
+export 'special_day_sheet.dart';
+export 'grid_logic.dart';
