+++ lib/pages/profile/profile_menu_card.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:bugaoshan/injection/injector.dart';
 import 'package:bugaoshan/l10n/app_localizations.dart';
 import 'package:bugaoshan/pages/about/about_page.dart';
-import 'package:bugaoshan/pages/course/course_schedule_setting.dart';
-import 'package:bugaoshan/pages/course/schedule_management_page.dart';
+import 'package:bugaoshan/pages/course/management/schedule_management_page.dart';
+import 'package:bugaoshan/pages/course/settings/course_schedule_setting.dart';
 import 'package:bugaoshan/pages/settings/software_setting_page.dart';
 import 'package:bugaoshan/providers/app_config_provider.dart';
 import 'package:bugaoshan/providers/course_provider.dart';
